将折扣改为1%~99%，用浮点数表示，但数据库中为100倍的整数

+++ BackEnd/src/main/java/com/example/tomatomall/po/Cart.java
@@ -30,5 +30,5 @@ public class Cart {
     private boolean ordered = false;
 
     @Column(name="discount")
-    private Integer discount;
+    private Float discount;
 }
