+++ BackEnd/src/main/java/com/example/tomatomall/vo/CartVO.java
@@ -16,6 +16,6 @@ public class CartVO {
     private String cover;
     private String detail;
     private Integer quantity;
-    private Integer discount;
+    private Float discount;
     private boolean ordered;
 }
